Added Slope calculations and modified stanleyInput accordingly.
