+++ ros2/.gitignore
@@ -1,5 +1,6 @@
 install/
 log/
+build
 */*/build
 */*/install
 */*/log
