+++ ros2/igvc_msgs/msg/StanleyInput.msg
@@ -1,0 +1,5 @@
+float32 CTE = 0.0
+float32 HE = 0.0
+bool LEFTLANE = False
+bool ERROR = False
+bool SLOPE = 0
