+++ ros2/lane_following/src/LaneDetection.py
@@ -8,139 +8,8 @@ from cv_bridge import CvBridge
 import cv2
 import numpy as np
 from std_msgs.msg import String
-
-class Individual_Follower:
-
-    def __init__(self):
-
-        self._fit = None
-        self._binary_warped = None
-
-    def set_binwarp(self, binwarp: np.ndarray):
-        self._binary_warped = binwarp
-
-    def Plot_Line(self, smoothen=False, prevFrameCount=6):
-        # Number of windows for sliding windows
-        nwindows = 9
-        # Set the width of the windows +/- margin
-        margin = 100
-        # Set minimum number of pixels found to recenter window
-        minpix = 20
-
-        histogram = np.sum(
-            self._binary_warped[self._binary_warped.shape[0] // 2 :, :], axis=0
-        )
-        # Create an output image to draw on and visualize the result
-        out_img = (
-            np.dstack(
-                (self._binary_warped, self._binary_warped, self._binary_warped)
-            )
-            * 255
-        )
-        base = np.argmax(histogram[:])
-
-        window_height = np.int32(self._binary_warped.shape[0] / nwindows)
-        # Create empty list to hold pixel indices
-        lane_inds = []
-
-        # Identify the x and y positions of all nonzero pixels in the image
-        nonzero = self._binary_warped.nonzero()
-        nonzeroy = np.array(nonzero[0])
-        nonzerox = np.array(nonzero[1])
-        current = base
-
-        # Used for calculating the number of "empty windows".
-        # Relevant for our lane determination: the more empty windows, the more likely this is the dashed line
-        empty_windows = 0
-        # Step through the windows one by one
-        for window in range(nwindows):
-            # Identify window boundaries in x and y (and right and left)
-            win_y_low = (
-                self._binary_warped.shape[0] - (window + 1) * window_height
-            )
-            win_y_high = self._binary_warped.shape[0] - window * window_height
-            win_x_low = current - margin
-            win_x_high = current + margin
-            # Draw the windows on the visualization image
-            cv2.rectangle(
-                out_img,
-                (win_x_low, win_y_low),
-                (win_x_high, win_y_high),
-                (0, 255, 0),
-                2,
-            )
-            # Identify the nonzero pixels in x and y within the window: inclusive of lower bounds
-            good_inds = (
-                (nonzeroy >= win_y_low)
-                & (nonzeroy < win_y_high)
-                & (nonzerox >= win_x_low)
-                & (nonzerox < win_x_high)
-            ).nonzero()[0]
-
-            # Append these indices to the lists
-            lane_inds.append(good_inds)
-            if len(good_inds) > minpix:
-                current = np.int32(np.mean(nonzerox[good_inds]))
-            else:
-                empty_windows += 1
-
-        # Concatenate the arrays of indices
-        if len(lane_inds) > 0:
-            lane_inds = np.concatenate(lane_inds)
-
-        # Extract line pixel positions
-        x_pos = nonzerox[lane_inds]
-        y_pos = nonzeroy[lane_inds]
-
-        # Fit a second order polynomial to each
-        if x_pos.any() and y_pos.any():
-            self._fit = np.polyfit(y_pos, x_pos, 2)
-        else:
-            return None, empty_windows, 0
-
-        ploty = np.linspace(
-            0, self._binary_warped.shape[0] - 1, self._binary_warped.shape[0]
-        )
-        fitx = np.polyval(self._fit, ploty)
-        out_img[nonzeroy[lane_inds], nonzerox[lane_inds]] = [255, 0, 0]
-
-        nonzero = self._binary_warped.nonzero()
-        nonzeroy = np.array(nonzero[0])
-        nonzerox = np.array(nonzero[1])
-
-        window_img = np.zeros_like(out_img)
-        # Generate a polygon to illustrate the search window area
-        # And recast the x and y points into usable format for cv2.fillPoly()
-        line_window1 = np.array(
-            [np.transpose(np.vstack([fitx - margin, ploty]))]
-        )
-        line_window2 = np.array(
-            [(np.transpose(np.vstack([fitx + margin, ploty])))]
-        )
-        line_pts = np.hstack((line_window1, line_window2))
-
-        # Draw the lane onto the warped blank imleft_bufferage
-        cv2.fillPoly(window_img, np.int_([line_pts]), color=(0, 255, 0))
-        line_pts = np.array(
-            [np.transpose(np.vstack([fitx, ploty]))], dtype=np.int32
-        )
-
-        cv2.polylines(
-            out_img, line_pts, isClosed=False, color=(0, 255, 255), thickness=3
-        )
-        # Evaluating heading error:
-
-        # This is the first window coordinates
-        y1 = self._binary_warped.shape[0] - (nwindows) * window_height
-        # This is the second window coordinates
-        y2 = self._binary_warped.shape[0] - (nwindows + 1) * window_height
-        # Returns polynomial values at a point
-        x1 = np.polyval(self._fit, y1)
-        x2 = np.polyval(self._fit, y2)
-
-        heading = math.atan((x2 - x1) / (y2 - y1)) * 180 / (math.pi)
-        result = cv2.addWeighted(out_img, 1, window_img, 0.3, 0)
-        return result, empty_windows, heading
+from igvc_msgs.msg import StanleyInput
+from Individual_Follower import Individual_Follower
 
 
 class Lane_Follower(Node):
@@ -171,6 +40,7 @@ class Lane_Follower(Node):
     MISSING_IMAGE_TOLERANCE = 100
     OVERFLOW = 1000.0
     FORMAT = (640, 480)
+    TOLERANCE = 100
 
     PIXELS_TO_METERS = 260.8269125
 
@@ -196,10 +66,13 @@ class Lane_Follower(Node):
         self.timer = self.create_timer(timer_period, self.timer_callback)
 
         # Publisher for error from the lane lines relative to the camera FOV
-        # 10 refers to the number of published references: standard value
-        self.crosstrack_pub = self.create_publisher(Float64, 'cross_track', 10)
-        self.lane_pub = self.create_publisher(String, "lane_state", 10)
-        self.heading_pub = self.create_publisher(Float64, "heading", 10)
+        # # 10 refers to the number of published references: standard value
+        # self.crosstrack_pub = self.create_publisher(Float64, 'cross_track', 10)
+        # self.lane_pub = self.create_publisher(String, "lane_state", 10)
+        # self.heading_pub = self.create_publisher(Float64, "heading", 10)
+
+        self.lane_error_pub = self.create_publisher(
+            StanleyInput, "stanley_data", 10)
 
         if Lane_Follower.GUI:
             self._bridge = CvBridge()
@@ -264,6 +137,7 @@ class Lane_Follower(Node):
         images = [(image_l, "left"), (image_r, "right")]
         left_heading = -1
         right_heading = -1
+        main_msg = StanleyInput()
         if not (success_l and success_r):
             return
 
@@ -301,20 +175,14 @@ class Lane_Follower(Node):
             self.img_publish("raw_" + image[1], frame)
             self.img_publish("tf_" + image[1], transformed_frame)
 
-        result_left, empty_left, left_heading = self._left_follower.Plot_Line()
-        result_right, empty_right, right_heading = (
-            self._right_follower.Plot_Line()
-        )
-        crosstrack = Float64()
+        result_left, empty_left, left_heading, left_slope = self._left_follower.Plot_Line()
+        result_right, empty_right, right_heading, right_slope = self._right_follower.Plot_Line()
         # TODO: Is this the behavior we want? Or do we need it to do something else if one of the lines is invalid?
         if result_left is not None or result_right is not None:
-            heading = Float64()
-            pos = self.measure_position_meters(result_left, result_right)
-            print("LEFT ERROR IS: " + str(empty_left))
-            print("RIGHT ERROR IS: " + str(empty_right))
-            crosstrack.data = pos
+            crosstrack = self.measure_position_meters(
+                result_left, result_right)
+            main_msg.CE = crosstrack
 
-            self.crosstrack_pub.publish(crosstrack)
             self._Left_Lane = (
                 True if empty_left < empty_right else self._Left_Lane
             )
@@ -322,25 +190,21 @@ class Lane_Follower(Node):
                 False if empty_left > empty_right else self._Left_Lane
             )
 
-            heading = Float64()
-            Lane = String()
-
             if self._Left_Lane:
-                Lane.data = "In Left Lane"
-                heading.data = left_heading
+                main_msg.LEFTLANE = True
+                main_msg.HE = left_heading
+                main_msg.SLOPE = left_slope
             else:
-                Lane.data = "In Right Lane"
-                heading.data = right_heading
-
-            self.lane_pub.publish(Lane)
-            self.heading_pub.publish(heading)
+                main_msg.LEFTLANE = False
+                main_msg.HE = right_heading
+                main_msg.SLOPE = right_slope
 
         else:
-            TOLERANCE = 100
             self._tolerance += 1
-            if self._tolerance > TOLERANCE:
-                crosstrack.data = Lane_Follower.OVERFLOW
-                self.crosstrack_pub.publish(crosstrack)
+            if self._tolerance > Lane_Follower.TOLERANCE:
+                main_msg.ERROR = True
+
+        self.lane_error_pub.publish(main_msg)
 
 
 def main(args=None):
